% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_metrics.R
\name{add_age_adjusted_rate}
\alias{add_age_adjusted_rate}
\title{Calculate Age-Adjusted Rate}
\usage{
add_age_adjusted_rate(
  total_burden,
  year,
  agr_by,
  pop.summary.dir = "../data/12_population_summary"
)
}
\arguments{
\item{total_burden}{A data frame containing the total burden information.}

\item{year}{A numeric value representing the year for which the calculation is to be done.}

\item{agr_by}{A character string specifying the aggregation level (e.g., "nation", "STATEFP", "county").}

\item{pop.summary.dir}{A character string specifying the directory path to the population summary data (default is "../data/12_population_summary").}
}
\value{
A data frame containing the calculated age-adjusted rate.
}
\description{
This function calculates the age-adjusted rate for a given total burden and population summary, using a standard population file.
}
\references{
\url{https://www.cdc.gov/nchs/data/nvsr/nvsr57/nvsr57_14.pdf}, page 125, Table VIII
}
