% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_metrics.R
\name{add_age_adjusted_rate_old}
\alias{add_age_adjusted_rate_old}
\title{Calculate Age-Adjusted Rate}
\usage{
add_age_adjusted_rate_old(
  total_burden,
  pop_summary,
  path_to_standartpopulation = "data/standartpopulation.xlsx"
)
}
\arguments{
\item{total_burden}{A data frame containing the total burden. Must contain columns "value", "min_age", and "max_age".}

\item{pop_summary}{A data frame containing the population summary. Must contain columns "min_age", "max_age", and "Population".}

\item{path_to_standartpopulation}{A string representing the path to the standard population Excel file.}
}
\value{
A data frame containing the age-adjusted rate.
}
\description{
This function calculates the age-adjusted rate for a given total burden and population summary, using a standard population file.
}
\references{
\url{https://www.cdc.gov/nchs/data/nvsr/nvsr57/nvsr57_14.pdf}, page 125, Table VIII
}
