% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_data.R
\name{add_column_if_not_exists}
\alias{add_column_if_not_exists}
\title{Add a column to a data frame if it doesn't exist}
\usage{
add_column_if_not_exists(df, col_name)
}
\arguments{
\item{df}{A data frame to which the column will be added.}

\item{col_name}{The name of the column to add.}
}
\value{
A data frame with the new column added if it didn't exist.
}
\description{
This function takes a data frame and a column name as arguments.
If the column doesn't exist in the data frame, it adds the column with a default value of "666".
}
