% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_metrics.R
\name{add_crude_rate}
\alias{add_crude_rate}
\title{Calculate Crude Rate}
\usage{
add_crude_rate(total_burden, pop_summary)
}
\arguments{
\item{total_burden}{A data frame containing the total burden. Must contain columns "value".}

\item{pop_summary}{A data frame containing the population summary. Must contain columns "Population".}
}
\value{
A data frame containing the crude rate.
}
\description{
This function calculates the crude rate for a given total burden and population summary.
}
