% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_data_managment.R
\name{add_custom_rows}
\alias{add_custom_rows}
\title{Add custom rows to a data frame}
\usage{
add_custom_rows(df)
}
\arguments{
\item{df}{A data frame that you want to add custom rows to.}
}
\value{
A data frame with the custom rows added.
}
\description{
This function takes a data frame \code{df} and adds custom rows to it by mutating specific columns.
The custom rows have the value "666" for the specified columns.
}
\examples{
# Create a sample data frame
sample_df <- data.frame(
  rural_urban_class = c("A", "B"),
  svi_bin = c(1, 2),
  svi_bin1 = c(3, 4),
  svi_bin2 = c(5, 6),
  svi_bin3 = c(7, 8),
  svi_bin4 = c(9, 10)
)

# Add custom rows
modified_df <- add_custom_rows(sample_df)
}
