% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_information.R
\name{add_rural_urban_class}
\alias{add_rural_urban_class}
\title{Add Rural-Urban Classification to a Dataframe}
\usage{
add_rural_urban_class(
  df,
  FIPS.code.column = "FIPS.code",
  Year.colum = "Year",
  path_to_rural_urban_class = "data/rural_urban_class.csv",
  silent = TRUE,
  corresponding_year = setNames(c(1990, rep(2010, 9), 2000, rep(2010, 8), 2000, rep(2010,
    7)), 1990:2016)
)
}
\arguments{
\item{df}{A dataframe that contains the data.}

\item{FIPS.code.column}{A string indicating the name of the column containing the FIPS code (default is "FIPS.code").}

\item{Year.colum}{A string indicating the name of the column containing the Year (default is "Year").}

\item{path_to_rural_urban_class}{A string with the path to the CSV file containing the rural-urban class lookup table (default is "data/rural_urban_class.csv").}
}
\value{
A dataframe with the added rural-urban classification.
}
\description{
This function takes a dataframe and specified columns for FIPS code and Year, then adds a rural-urban classification
using a lookup table provided in a CSV file.
}
