% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{calculate_mean_and_ci}
\alias{calculate_mean_and_ci}
\title{Calculate Mean and 95\% Confidence Interval}
\usage{
calculate_mean_and_ci(x)
}
\arguments{
\item{x}{A numeric vector for which the mean and 95\% CI are to be calculated.}
}
\value{
A data frame containing the mean, lower, and upper estimates of the 95\% CI.
}
\description{
This function takes a numeric vector and calculates the mean, lower, and upper
estimates of the 95\% confidence interval for the mean.
}
\examples{
calculate_mean_and_ci(c(1, 2, 3, 4, 5))
}
