% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{calculate_weighted_mean_ci}
\alias{calculate_weighted_mean_ci}
\title{Calculate Weighted Mean and Confidence Intervals}
\usage{
calculate_weighted_mean_ci(pm, pop_size, conf_level = 0.95, R = 10)
}
\arguments{
\item{pm}{A numeric vector representing the values for which the weighted mean is calculated.}

\item{pop_size}{A numeric vector representing the weights for the weighted mean.}

\item{conf_level}{The confidence level for the interval. Default is 0.95.}

\item{R}{The number of bootstrap replicates. Default is 10.}
}
\value{
A list containing the weighted mean and its lower and upper confidence intervals.
}
\description{
This function calculates the weighted mean and its confidence intervals.
}
\examples{
\dontrun{
pm <- c(1, 2, 3, 4)
pop_size <- c(10, 20, 30, 40)
result <- calculate_weighted_mean_ci(pm, pop_size)
}
}
