% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{can_be_numeric}
\alias{can_be_numeric}
\title{Check if values can be converted to numeric}
\usage{
can_be_numeric(x)
}
\arguments{
\item{x}{A vector of any type.}
}
\value{
A boolean indicating whether all elements can be converted to numeric or not.
}
\description{
This function checks if all values in a vector can be converted to numeric using as.numeric().
If there's at least one value that cannot be converted, it returns FALSE.
If all values can be converted, it returns TRUE.
}
\examples{
can_be_numeric(c("1", "2", "3")) # Returns: TRUE
can_be_numeric(c("1", "2", "b")) # Returns: FALSE
}
