% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers2.R
\name{check_package_usage}
\alias{check_package_usage}
\title{Check if a Package is Used in an R Script}
\usage{
check_package_usage(path_to_script, pkg)
}
\arguments{
\item{path_to_script}{A string providing the path to the R script.}

\item{pkg}{A string indicating the name of the package to check.}
}
\value{
A character vector with the names of the functions from the package used in the script,
or NULL if none are used.
}
\description{
This function checks if functions from a specified package are used in the given R script.
It returns the names of functions from the package that are used in the script, or NULL if none are used.
Additionally, a message indicating the usage status will be displayed.
}
\examples{
\dontrun{
result <- check_package_usage("path_to_your_script.R", "dplyr")
cat(result)
}
}
