% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_columns_to_numeric}
\alias{convert_columns_to_numeric}
\title{Convert columns of a dataframe that can be numeric to numeric}
\usage{
convert_columns_to_numeric(df)
}
\arguments{
\item{df}{A dataframe with the columns to convert.}
}
\value{
A dataframe with columns converted to numeric type if possible.
}
\description{
This function takes a dataframe as input and iterates over each column.
If a column can be converted to a numeric type, it performs the conversion.
}
\examples{
# create a sample data frame
df <- data.frame(a = c("1", "2", "3"), b = c("a", "b", "c"))
df <- convert_columns_to_numeric(df)

}
