% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_combined_plot}
\alias{create_combined_plot}
\title{Create Combined Plot}
\usage{
create_combined_plot(
  plots,
  legend_plot,
  row_annotations,
  column_annotations,
  y_axis = "Age-adjusted mortality per 100,000"
)
}
\arguments{
\item{plots}{A list of ggplot objects to be combined.}

\item{legend_plot}{A ggplot object for the legend.}

\item{row_annotations}{A list of strings for row annotations.}

\item{column_annotations}{A list of strings for column annotations.}

\item{y_axis}{A string for the y-axis label.}
}
\value{
A combined plot.
}
\description{
This function combines multiple plots and additional elements like text and legends into a single plot.
}
\examples{
# Assuming `plots`, `legend_plot`, `row_annotations`, and `column_annotations` are defined
create_combined_plot(plots, legend_plot, row_annotations, column_annotations, y_axis = "Age-adjusted mortality per 100,000")
}
