% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_directory}
\alias{create_directory}
\title{Create directory if it does not exist}
\usage{
create_directory(dir_path)
}
\arguments{
\item{dir_path}{A character string specifying the path to the directory to be created.}
}
\value{
Invisible NULL. The function is called for its side effect of creating a directory.
}
\description{
This function checks if a directory at the specified path exists.
If it does not exist, the function creates it.
}
\examples{
\dontrun{
create_directory("path/to/your/directory")
}
}
