% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{delta_method_rel_diff}
\alias{delta_method_rel_diff}
\title{Estimate the Mean and Confidence Interval for the Relative Difference Using the Delta Method}
\usage{
delta_method_rel_diff(mean_x, lb_x, ub_x, mean_y, lb_y, ub_y, alpha = 0.05)
}
\arguments{
\item{mean_x}{Mean of the first random variable X}

\item{lb_x}{Lower bound of the confidence interval for X}

\item{ub_x}{Upper bound of the confidence interval for X}

\item{mean_y}{Mean of the second random variable Y}

\item{lb_y}{Lower bound of the confidence interval for Y}

\item{ub_y}{Upper bound of the confidence interval for Y}

\item{alpha}{Significance level for the confidence interval (default is 0.05 for a 95\% CI)}
}
\value{
A list containing the estimated mean, lower bound, and upper bound of the confidence interval for g(X, Y)
}
\description{
Estimate the Mean and Confidence Interval for the Relative Difference Using the Delta Method
}
\details{
The Delta Method is used to approximate the variance of a function g(X, Y) of random variables X and Y.
In this case, g(X, Y) = (X - Y) / X.
}
