% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{delta_method_weighted_avg}
\alias{delta_method_weighted_avg}
\title{Estimate Point Estimate and Confidence Interval of Weighted Average Using Delta Method}
\usage{
delta_method_weighted_avg(pm, pm_lower, pm_upper, pop_size)
}
\arguments{
\item{pm}{A numeric vector of point estimates for the random variable X.}

\item{pm_lower}{A numeric vector of lower bounds of the confidence interval for X.}

\item{pm_upper}{A numeric vector of upper bounds of the confidence interval for X.}

\item{pop_size}{A numeric vector of population sizes.}
}
\value{
A list containing the point estimate and confidence interval of the weighted average.
}
\description{
Estimate Point Estimate and Confidence Interval of Weighted Average Using Delta Method
}
\examples{
pm <- c(0.2, 0.3, 0.4)
pm_lower <- c(0.1, 0.2, 0.3)
pm_upper <- c(0.3, 0.4, 0.5)
pop_size <- c(100, 200, 300)
delta_method_weighted_avg(pm, pm_lower, pm_upper, pop_size)
}
