% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{diagnose_join_issues}
\alias{diagnose_join_issues}
\title{Diagnose join issues between two data frames}
\usage{
diagnose_join_issues(df1, df2, join_cols)
}
\arguments{
\item{df1}{A data frame to be joined}

\item{df2}{Another data frame to be joined}

\item{join_cols}{A character vector specifying the columns to join on}
}
\value{
A data frame after performing the required joins
}
\description{
Diagnose join issues between two data frames
}
\examples{
diagnose_join_issues(df1, df2, c("col1", "col2"))
}
