% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{find_na_rows}
\alias{find_na_rows}
\title{Find Rows with NA Values in a Dataframe}
\usage{
find_na_rows(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with rows containing at least one NA value.
}
\description{
This function returns rows with at least one NA value from a provided dataframe.
}
\examples{
df <- data.frame(a = c(1, 2, NA, 4), b = c(NA, 6, 7, 8))
find_na_rows(df)
}
