% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_data_managment.R
\name{generate_filtered_dfs}
\alias{generate_filtered_dfs}
\title{Generate Filtered Dataframes Based on Condition Combinations}
\usage{
generate_filtered_dfs(df)
}
\arguments{
\item{df}{A dataframe to filter.}
}
\value{
A list of filtered dataframes.
}
\description{
This function takes a dataframe and filters it based on all possible combinations
of the conditions specified for each variable.
}
