% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_group_colors}
\alias{get_group_colors}
\title{Get Group Colors}
\usage{
get_group_colors(df = NULL)
}
\arguments{
\item{df}{A data frame containing the groups. Default is NULL.}
}
\value{
A named vector of colors for the groups.
}
\description{
This function returns a named vector of colors for various demographic and social groups.
If a data frame is provided, the function will filter the colors to only include groups present in the data frame.
}
\examples{
get_group_colors()
get_group_colors(df = data.frame(Column1 = c("NH White", "Hispanic or Latino White"), Column2 = c("Black American", "White")))
}
