% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{group_summarize_add_column}
\alias{group_summarize_add_column}
\title{Group, Summarize and Add New Column}
\usage{
group_summarize_add_column(df, column, new_col_value)
}
\arguments{
\item{df}{A dataframe}

\item{column}{A character string specifying the column name to exclude from grouping and to add}

\item{new_col_value}{A character string specifying the value to fill the new column}
}
\value{
A dataframe summarized and with the new column added.
}
\description{
This function groups a dataframe by all columns except the specified one,
sums the 'Deaths' column, and adds a new column with a provided name and value.
}
\examples{
df <- data.frame(a = c(1, 2, 3, 4), b = c(5, 6, 7, 8), Deaths = c(10, 20, 30, 40))
group_summarize_add_column(df, column = "b", new_col_value = "A")
}
