% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_data.R
\name{has_overlaps}
\alias{has_overlaps}
\title{Check if the given data.frame has overlapping age intervals}
\usage{
has_overlaps(df)
}
\arguments{
\item{df}{A data.frame with columns "min_age" and "max_age"}
}
\value{
A boolean value indicating if the data.frame has overlapping age intervals.
}
\description{
Check if the given data.frame has overlapping age intervals
}
\examples{
df <- data.frame(min_age = c(25, 26), max_age = c(25, 26))
has_overlaps(df)
}
