% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inner_join_age}
\alias{inner_join_age}
\title{Inner join two data frames and filter by age.}
\usage{
inner_join_age(df1, df2, by, right_outer = TRUE)
}
\arguments{
\item{df1}{The first data frame to join.}

\item{df2}{The second data frame to join.}

\item{by}{A vector of variable names to join by.}

\item{right_outer}{A logical value. If TRUE, the function performs a right outer join.}
}
\value{
A data frame that is the result of joining df1 and df2 and filtered by min_age and max_age.
}
\description{
This function performs an inner join on two data frames, filters based on minimum and maximum age and
removes redundant age columns. It allows for the option of specifying the joining type to be a right outer join.
}
\examples{
df1 <- data.frame(id = c(1, 2, 3), min_age = c(20, 25, 30), max_age = c(40, 45, 50))
df2 <- data.frame(id = c(1, 2, 4), min_age = c(30, 35, 40), max_age = c(50, 55, 60))
inner_join_age(df1, df2, by = "id", right_outer = TRUE)
}
