% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inner_join_age_right_outer}
\alias{inner_join_age_right_outer}
\title{Inner join two data frames using right outer join and filter by age.}
\usage{
inner_join_age_right_outer(df1, df2, by, group_column = NULL)
}
\arguments{
\item{df1}{The first data frame to join.}

\item{df2}{The second data frame to join.}

\item{by}{A vector of variable names to join by.}
}
\value{
A data frame that is the result of joining df1 and df2 using a right outer join and filtered by min_age and max_age.
}
\description{
This function is a wrapper for the \code{\link{inner_join_age}} function with the \code{right_outer} parameter set to TRUE.
}
\examples{
df1 <- data.frame(id = c(1, 2, 3), min_age = c(20, 25, 30), max_age = c(40, 45, 50))
df2 <- data.frame(id = c(1, 2, 4), min_age = c(30, 35, 40), max_age = c(50, 55, 60))
inner_join_age_right_outer(df1, df2, by = "id")
}
