% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{insert_line_break}
\alias{insert_line_break}
\title{Insert a line break at an appropriate space in a string}
\usage{
insert_line_break(string, n = 20)
}
\arguments{
\item{string}{Character vector containing the string to be modified.}

\item{n}{Integer specifying the length threshold for line break insertion.}
}
\value{
A modified string with a line break at an appropriate space if the original string's length is n or more.
}
\description{
This function takes a string and an integer n as input.
If the string has length n or more, it inserts a "\n" at a space character,
ideally as centrally as possible within the string.
}
\examples{
insert_line_break("alle meine Entchen schwimmen auf dem See", 20)
insert_line_break("short string", 20)
}
