% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_data.R
\name{is_partition}
\alias{is_partition}
\title{Check if the given data.frame represents a partition of an interval}
\usage{
is_partition(df)
}
\arguments{
\item{df}{A data.frame with columns "min_age" and "max_age"}
}
\value{
A boolean value indicating if the data.frame represents a partition.
}
\description{
Check if the given data.frame represents a partition of an interval
}
\examples{
df <- data.frame(min_age = c(25, 30, 35, 40), max_age = c(29, 34, 39, 44))
is_partition(df)
}
