% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_code_handling.R
\name{outer_difference_heatmap}
\alias{outer_difference_heatmap}
\title{Calculate the Difference Matrix}
\usage{
outer_difference_heatmap(list_all_counties)
}
\arguments{
\item{list_all_counties}{A list containing information about all counties.}
}
\value{
A heatmap of the difference matrix.
}
\description{
This function calculates a difference matrix using an extended version of the \code{outer} function.
It applies a specific function to calculate the differences between counties and returns a heatmap.
}
