% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{parametric_bootstrap_stat}
\alias{parametric_bootstrap_stat}
\title{Parametric Bootstrap for Custom Statistic}
\usage{
parametric_bootstrap_stat(data, stat_function, n_boot = 1000)
}
\arguments{
\item{data}{A data frame containing columns 'mean', 'lower', and 'upper', which represent
the point estimates and the lower and upper bounds of the 95\% confidence intervals for each X_i.}

\item{stat_function}{A function to calculate the statistic of interest from a sample.}

\item{n_boot}{Number of bootstrap samples (default is 1000).}
}
\value{
A list containing the mean and 95\% confidence intervals (lower and upper) for the custom statistic.
}
\description{
This function performs parametric bootstrapping to estimate a custom statistic and its 95\% confidence intervals,
taking into account existing confidence intervals for each X_i.
}
\examples{
original_data <- data.frame(mean = c(1, 2, 3, 4, 5), lower = c(0.5, 1.5, 2.5, 3.5, 4.5), upper = c(1.5, 2.5, 3.5, 4.5, 5.5))
cov_function <- function(x) { sqrt(var(x)) / mean(x) }
parametric_bootstrap_stat(original_data, stat_function = cov_function)
}
