% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_attr_all_burd}
\alias{plot_attr_all_burd}
\title{Plot Combined Data for attr_burd and all_burd}
\usage{
plot_attr_all_burd(attr_burd, all_burd, color.column, split.column)
}
\arguments{
\item{attr_burd}{Data frame containing the attr_burd data.}

\item{all_burd}{Data frame containing the all_burd data.}

\item{color.column}{The name of the column to be used for coloring the plots.}

\item{split.column}{The name of the column to be used for splitting the data frames.}
}
\value{
A combined ggplot object.
}
\description{
This function takes two data frames, attr_burd and all_burd, along with the names of
the columns to be used for coloring and splitting. It then generates a combined plot.
}
\examples{
\dontrun{
result <- plot_attr_all_burd(attr_burd, all_burd, "Ethnicity", "svi_bin1")
}
}
