% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_df}
\alias{plot_df}
\title{Plot Data Frame with ggplot2}
\usage{
plot_df(df, color.column, group.colors = get_group_colors(df))
}
\arguments{
\item{df}{Data frame containing the data to be plotted.}

\item{color.column}{Character string specifying the column to be used for coloring.}

\item{group.colors}{Vector of colors to be used for the groups. Default is obtained from \code{get_group_colors()}.}
}
\value{
A ggplot object.
}
\description{
This function takes a data frame and plots it using ggplot2. It allows for optional
lower and upper bounds to be plotted as a ribbon.
}
\examples{
plot_df(df, "Ethnicity")
}
