% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{power_set}
\alias{power_set}
\title{Generate the power set of a vector}
\usage{
power_set(x)
}
\arguments{
\item{x}{A vector whose power set you want to generate.}
}
\value{
A list containing all subsets of the input vector.
}
\description{
Generate the power set of a vector
}
