% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_data}
\alias{read_data}
\title{Read Data from CSV file}
\usage{
read_data(path)
}
\arguments{
\item{path}{A string. The path to the CSV file to read.}
}
\value{
A data frame with columns: Year, min_age, max_age, county, rural_urban_class,
Deaths, Education, Race. The "Race" column is read as a factor with specific levels.
}
\description{
This function reads data from a CSV file at a given path. It specifies column types
for the data to ensure it is read correctly. This function uses the readr package
to read the CSV.
}
\examples{
df <- read_data("path/to/your/file.csv")

}
