% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{replace_missing_vectors}
\alias{replace_missing_vectors}
\title{Replace missing values in a vector}
\usage{
replace_missing_vectors(vec1, vec2, is_missing, vec_name = NULL)
}
\arguments{
\item{vec1}{A vector in which missing values should be replaced.}

\item{vec2}{A vector from which replacement values should be taken.}

\item{is_missing}{A function that identifies missing values.}

\item{vec_name}{An optional name for the vectors, used in printed messages. Default is NULL.}
}
\value{
A vector with missing values replaced where possible.
}
\description{
This function replaces missing values in the first vector (\code{vec1}) with corresponding values from a second vector (\code{vec2}),
if the corresponding values are not missing.
If no missing values are detected in \code{vec1}, the function returns \code{vec1} without performing any further operations.
}
\examples{
vec1 <- c(1, NA, 3, NA, 5)
vec2 <- c(2, 4, 6, 8, 10)
replace_missing_vectors(vec1, vec2, is.na)
}
