% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{replace_values}
\alias{replace_values}
\title{Replace specific values in a dataframe}
\usage{
replace_values(df, findreplace, NA_string = "Unique_NA_String", silent = TRUE)
}
\arguments{
\item{df}{A dataframe with the columns to replace values in.}

\item{findreplace}{A dataframe defining what values to find and what to replace them with.}

\item{NA_string}{A unique string to replace NA values with during processing (default is "Unique_NA_String"). Should not be present in df.}
}
\value{
A dataframe with replaced values as defined by the findreplace dataframe.
}
\description{
This function takes a dataframe and a findreplace dataframe as input. It replaces specific values in the dataframe
as defined by the findreplace dataframe.
}
\examples{
# create a sample data frame
df <- data.frame(a = c("1", "2", "3"), b = c("a", "b", "c"))
findreplace <- data.frame(replacecolumns = c("a", "b"), from = c("1", "a"), to = c("10", "z"))
df <- replace_values(df, findreplace)

}
