% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{run_script}
\alias{run_script}
\title{Run script based on system type}
\usage{
run_script(script, args = "")
}
\arguments{
\item{script}{A string specifying the name/path of the script to be run.}

\item{args}{A string specifying the arguments to be passed to the script. Default is "" (no arguments).}
}
\value{
NULL. This function is mainly called for its side effects (running the script and printing the results).
}
\description{
This function runs a given R script file with optional arguments, using different settings depending on the current system's type.
The function will also print which script was run and how long it took.
}
\examples{
\dontrun{
run_script("my_script.R")
run_script("my_script.R", "--arg value")
}
}
