% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_data_managment.R
\name{search_files}
\alias{search_files}
\title{Search for Files in a Directory}
\usage{
search_files(directory, string, ends_with = FALSE)
}
\arguments{
\item{directory}{A string specifying the path to the directory.}

\item{string}{The string to search for in the file and directory names.}

\item{ends_with}{A logical value. If TRUE, only files and directories whose names
end with the specified string are returned. If FALSE (the default), files and directories
whose names contain the string anywhere are returned.}
}
\value{
A character vector containing the full paths to the matching files and directories.
}
\description{
This function takes a directory path and a string, and returns the path to all files
and directories in that path, where the name contains the string.
Optionally, it can also return only those files and directories whose names end with the string.
}
\examples{
\dontrun{
search_files("path/to/directory", "test")
search_files("path/to/directory", "test", ends_with = TRUE)
}
}
