% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{simplify_columns_df}
\alias{simplify_columns_df}
\title{Simplify Data Frame Columns}
\usage{
simplify_columns_df(df)
}
\arguments{
\item{df}{A data frame containing the columns to be simplified.}
}
\value{
A data frame with character columns that have been simplified based on the described rules.
}
\description{
This function iterates through the columns of the provided data frame,
attempting to convert character columns to numeric if possible, and then back to character.
If the column is a factor, it will be converted to a character column.
}
\examples{
df <- data.frame(a = c("1", "2", "3"), b = as.factor(c("a", "b", "c")))
df <- simplify_columns_df(df)

}
