% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_column_by_group}
\alias{summarize_column_by_group}
\title{Summarize a specified column by grouping other columns}
\usage{
summarize_column_by_group(df, group_column)
}
\arguments{
\item{df}{A dataframe to be summarized.}

\item{group_column}{The column name to be summarized (character).}
}
\value{
A dataframe with the summarized column.
}
\description{
This function takes a dataframe and a column name to summarize.
It groups by all other columns and then summarizes the specified column.
If the specified column is not in the dataframe, it throws an error.
}
\examples{
df3 <- data.frame(A = c('a', 'a', 'b', 'b'), B = c('x', 'x', 'y', 'y'), value = c(1,2,3,4))
group_column <- "value"
summarize_column_by_group(df3, group_column)
}
