% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_information.R
\name{add_social_vuln_index}
\alias{add_social_vuln_index}
\title{Add Social Vulnerability Index (SVI) to a Dataframe}
\usage{
add_social_vuln_index(
  df,
  FIPS.code.column = "FIPS.code",
  Year.colum = "Year",
  path_to_svi = "data/08_svi/svi_lookup_table.csv",
  silent = TRUE
)
}
\arguments{
\item{df}{A dataframe that contains the data.}

\item{FIPS.code.column}{A string indicating the name of the column containing the FIPS code (default is "FIPS.code").}

\item{path_to_svi}{A string with the path to the CSV file containing the SVI lookup table (default is "data/08_svi/svi_lookup_table.csv").}
}
\value{
A dataframe with the added SVI values.
}
\description{
This function takes a dataframe and a specified FIPS code column, then adds a Social Vulnerability Index (SVI)
using a lookup table provided in a CSV file.
}
\examples{
# Assuming existence of a dataframe 'df' with a "FIPS.code" column
df_with_svi <- add_social_vuln_index(df)

}
