% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_used.R
\name{check_package_usage}
\alias{check_package_usage}
\title{Check Package Usage in an R Script}
\usage{
check_package_usage(path_to_script, pkgs)
}
\arguments{
\item{path_to_script}{A string representing the path to the R script.}

\item{pkgs}{A character vector containing the names of the packages to check.}
}
\value{
A character vector containing the names of the packages that are used in the script, or NULL if none of the specified packages are used.
}
\description{
This function checks which functions from a given set of packages are used in an R script.
It takes the path to the R script and a vector of package names, then returns the packages that are actually used.
}
\examples{
\dontrun{
  path <- "path/to/your/script.R"
  packages <- c("dplyr", "ggplot2", "tidyr")
  result <- check_package_usage(path, packages)
  print(result)
}
}
