% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_df}
\alias{plot_df}
\title{Plot data frame using ggplot2}
\usage{
plot_df(
  df,
  color.column = NULL,
  group.colors = NULL,
  remove_legend = TRUE,
  remove_y_titel = TRUE,
  title = NA
)
}
\arguments{
\item{df}{Data frame containing the data to plot.}

\item{color.column}{The column in df that will be used to color the lines.}

\item{group.colors}{A vector of colors to use for the different groups.}

\item{remove_legend}{Logical value to indicate whether to remove the legend or not.}

\item{remove_y_titel}{Logical value to indicate whether to remove the y-axis title or not.}

\item{title}{Title for the plot.}
}
\value{
A ggplot object.
}
\description{
Plot data frame using ggplot2
}
\examples{
plot_df(data.frame(Year = c(2000, 2001), mean = c(1, 2), color = c("red", "blue")))
}
