% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{update_ylim}
\alias{update_ylim}
\title{Update y-axis limits for a list of ggplot objects}
\usage{
update_ylim(ggplot_list)
}
\arguments{
\item{ggplot_list}{A list of ggplot objects.}
}
\value{
A list of ggplot objects with updated y-axis limits.
}
\description{
This function takes a list of ggplot objects and sets the y-axis limits
to be the same for all plots, based on the maximum value in either the 'upper' or 'value' column.
Throws an error if neither column is present.
}
\examples{
g1 <- ggplot(data.frame(x = 1:10, upper = rnorm(10)), aes(x, upper)) + geom_line()
g2 <- ggplot(data.frame(x = 1:10, value = rnorm(10)), aes(x, value)) + geom_line()
g3 <- ggplot(data.frame(x = 1:10, upper = rnorm(10)), aes(x, upper)) + geom_line()
ggplot_list <- list(g1, g2, g3)
updated_ggplot_list <- update_ylim(ggplot_list)
}
